refactor: move local utils
