+++ src/shared/components/Drawer/components/DialogSearchAll/components/TableDialogSearchAll/components/GetActions/utils/getSubDirectory/types/get-sub-directory.types.ts
